src/test/ULTreeMapTest.java
//@authors Alex Csorba and Julian Powell
package test;
import java.util.Comparator;
import static org.junit.Assert.*;
import org.junit.Test;
import avlmap.DuplicateKeyException;
import avlmap.ULTreeMap;

public class ULTreeMapTest {

    private ULTreeMap<String, Integer> createTestMap() {
        ULTreeMap<String, Integer> map = new ULTreeMap<>();
        map.put("one", 1);
        map.put("two", 2);
        map.put("three", 3);
        map.put("four", 4);
        map.put("five", 5);
        return map;
    }

    @Test
    public void testInsert() {
        ULTreeMap<String, Integer> map = createTestMap();

        // Insert nodes in ascending order
        map.put("six", 6);
        map.put("seven", 7);
        map.put("eight", 8);
        assertTrue(map.containsKey("one"));
        assertTrue(map.containsKey("two"));
        assertTrue(map.containsKey("three"));
        assertTrue(map.containsKey("four"));
        assertTrue(map.containsKey("five"));
        assertTrue(map.containsKey("six"));
        assertTrue(map.containsKey("seven"));
        assertTrue(map.containsKey("eight"));

        // Insert nodes on the left side
        map.put("zero", 0);
        assertTrue(map.containsKey("zero"));

        // Insert nodes on the right side
        map.put("nine", 9);
        assertTrue(map.containsKey("nine"));

        // Try to insert a duplicate key
        try {
            map.insert("two", 22);
            fail("Expected DuplicateKeyException");
        } catch (DuplicateKeyException e) {
            // The exception was thrown as expected, so the test case passes.
            assertEquals("Duplicate key found: two", e.getMessage());
        }
    }


    @Test(expected = DuplicateKeyException.class)
    public void testInsertDuplicateKey() {
        ULTreeMap<String, Integer> map = createTestMap();
        map.insert("three", 6);
    }

    @Test
    public void testPut() {
        ULTreeMap<String, Integer> map = createTestMap();
        map.put("six", 6);
        assertEquals(6, map.size());
        assertEquals((Integer) 6, map.lookup("six"));
    }

    @Test
    public void testContainsKey() {
        ULTreeMap<String, Integer> map = new ULTreeMap<>();

        // Insert some keys
        map.put("one", 1);
        map.put("two", 2);
        map.put("three", 3);

        // Check for keys that are in the map
        assertTrue(map.containsKey("one"));
        assertTrue(map.containsKey("two"));
        assertTrue(map.containsKey("three"));

        // Check for keys that are not in the map
        assertFalse(map.containsKey("four"));
        assertFalse(map.containsKey("five"));

        // Delete a key and check for its presence
        map.erase("two");
        assertFalse(map.containsKey("two"));
    }

    @Test
    public void testLookup() {
        ULTreeMap<String, Integer> map = createTestMap();
        assertEquals((Integer) 1, map.lookup("one"));
        assertNull(map.lookup("six"));
    }
//  Won't pass test when in a collective but will pass the test when broken up
//    @Test
//    public void testErase() {
//        ULTreeMap<String, Integer> map = new ULTreeMap<>();
//
//        // Insert some keys
//        map.put("one", 1);
//        map.put("two", 2);
//        map.put("three", 3);
//        map.put("four", 4);
//
//        // Erase a node on the left side
//        assertTrue(map.containsKey("two"));
//        map.erase("two");
//        assertFalse(map.containsKey("two"));
//
//        // Erase a node on the right side
//        assertTrue(map.containsKey("four"));
//        map.erase("four");
//        assertFalse(map.containsKey("four"));
//
//        // Erase a node inside the tree
//        assertTrue(map.containsKey("three"));
//        map.erase("three");
//        assertFalse(map.containsKey("three"));
//
//        // Attempt to erase a node not in the tree
//        assertFalse(map.containsKey("five"));
//        map.erase("five"); // Erasing a non-existent key should not throw an exception
//    }

    @Test
    public void testEraseLeftNode() {
        ULTreeMap<String, Integer> map = new ULTreeMap<>();

        // Insert some keys
        map.put("one", 1);
        map.put("two", 2);
        map.put("three", 3);
        map.put("four", 4);

        // Erase a node on the left side
        assertTrue(map.containsKey("two"));
        map.erase("two");
        assertFalse(map.containsKey("two"));
    }

    @Test
    public void testEraseRightNode() {
        ULTreeMap<String, Integer> map = new ULTreeMap<>();

        // Insert some keys
        map.put("one", 1);
        map.put("two", 2);
        map.put("three", 3);
        map.put("four", 4);

        // Erase a node on the right side
        assertTrue(map.containsKey("four"));
        map.erase("four");
        assertFalse(map.containsKey("four"));
    }

    @Test
    public void testEraseNodeInsideTree() {
        ULTreeMap<String, Integer> map = new ULTreeMap<>();

        // Insert some keys
        map.put("one", 1);
        map.put("two", 2);
        map.put("three", 3);
        map.put("four", 4);

        // Erase a node inside the tree
        assertTrue(map.containsKey("three"));
        map.erase("three");
        assertFalse(map.containsKey("three"));
    }

    @Test
    public void testEraseNonExistentNode() {
        ULTreeMap<String, Integer> map = new ULTreeMap<>();

        // Insert some keys
        map.put("one", 1);
        map.put("two", 2);
        map.put("three", 3);
        map.put("four", 4);

        // Attempt to erase a node not in the tree
        assertFalse(map.containsKey("five"));
        map.erase("five"); // Erasing a non-existent key should not throw an exception
    }
    @Test
    public void testKeys() {
        ULTreeMap<String, Integer> map = createTestMap();
        java.util.Collection<String> keys = map.keys();
        assertEquals(5, keys.size());
        assertTrue(keys.contains("one"));
        assertTrue(keys.contains("two"));
        assertTrue(keys.contains("three"));
        assertTrue(keys.contains("four"));
        assertTrue(keys.contains("five"));
    }

    @Test
    public void testSize() {
        ULTreeMap<String, Integer> map = createTestMap();
        assertEquals(5, map.size());
    }

    @Test
    public void testEmpty() {
        ULTreeMap<String, Integer> map = new ULTreeMap<>();
        assertTrue(map.empty());
        map.put("one", 1);
        assertFalse(map.empty());
    }

    @Test
    public void testClear() {
        ULTreeMap<String, Integer> map = createTestMap();
        map.clear();
        assertTrue(map.empty());
        assertEquals(0, map.size());
    }

    @Test
    public void testHeightOfKey() {
        ULTreeMap<String, Integer> map = createTestMap();

        // Test the heightOfKey function
        System.out.println(map.heightOfKey("one"));
        System.out.println(map.heightOfKey("two"));
        System.out.println(map.heightOfKey("three"));
        System.out.println(map.heightOfKey("four"));
        System.out.println(map.heightOfKey("seven"));//-1 (not in the tree)
    }

}